+++ jjava-kernel/src/main/java/org/dflib/jjava/kernel/execution/CodeEvaluator.java
@@ -89,7 +89,7 @@ public class CodeEvaluator {
 
     protected Object evalSingle(JShell shell, String code) {
 
-        List<SnippetEvent> events = shell.eval(code);
+        List<SnippetEvent> events = shell.eval(code.strip());
 
         Object result = null;