+++ jjava-kernel/src/main/java/org/dflib/jjava/kernel/execution/JJavaExecutionControl.java
@@ -76,11 +76,6 @@ class JJavaExecutionControl extends DirectExecutionControl {
         executor.shutdownNow();
     }
 
-    @Override
-    public void redefine(ClassBytecodes[] cbcs) {
-        loaderDelegate.classesRedefined(cbcs);
-    }
-
     public Object takeResult(String id) {
         Object result = this.results.remove(id);
         if (result == null) {